Day 1.
to start i made a way to login and a way to make an account.  when you create an account you either make a admin account ot a user account. if you sign in with an admin account it will take you to the admin page where in the future you will be able to delete comments and or users. if you login with a user account it will take you to the user page where in the future you will be able to post comments.