more user stuff
i added a way for users to add comments and delete them. i am trying to add a way for the user to edit their comments but so far i get errors that i cant fix. i will work on that tomorrw but so far i think this project is coming together and i am excited to see how it unfolds.